pyproject.toml
[project]
name = "simeg-order-entry-prototype"
version = "0.1.0"
description = "AI-powered order entry prototype for Simeg"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "google-genai>=0.2.0",
    "python-multipart>=0.0.6",
    "gradio>=4.16.0",
    "pandas>=2.1.0",
    "requests>=2.31.0",
    "pytest>=7.4.0",
    "hypothesis>=6.92.0",
    "openpyxl"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["BE", "FE"]
